+++ src/documentDecorationManager.ts
@@ -54,7 +54,9 @@ export default class DocumentDecorationManager {
     public onDidChangeSelection(event: TextEditorSelectionChangeEvent) {
         const documentDecoration = this.getDocumentDecorations(event.textEditor.document);
         if (documentDecoration &&
-            (documentDecoration.settings.highlightActiveScope || documentDecoration.settings.showBracketsInGutter)) {
+            (documentDecoration.settings.highlightActiveScope ||
+                documentDecoration.settings.showBracketsInGutter ||
+                documentDecoration.settings.showScopeLine)) {
             documentDecoration.updateScopeDecorations(event);
         }
     }